Add LogOutTest, PersonalAccountTest
