+++ src/test/java/OpenLoginPageTest.java
@@ -3,7 +3,7 @@ import org.junit.Test;
 
 import static com.codeborne.selenide.Selenide.open;
 
-public class OpenLoginPage {
+public class OpenLoginPageTest {
     HomePage homePage = open(HomePage.pageURL, HomePage.class);
     @Test
     //Проверь переход по клику на «Личный кабинет»
